WriteDirChanger has been moved to hpp/cpp
