+++ storage/storage_integration_tests/write_dir_changer.hpp
@@ -1,0 +1,12 @@
+#include "std/string.hpp"
+
+class WritableDirChanger
+{
+public:
+  WritableDirChanger(string testDir);
+  ~WritableDirChanger();
+
+private:
+  string const m_writableDirBeforeTest;
+  string const m_testDirFullPath;
+};
